requirements.txt
aiogram~=2.11.2
aiohttp
aioredis
python-dotenv~=0.15.0
numpy~=1.19.3

requests~=2.25.0
beautifulsoup4~=4.9.3
fuzzywuzzy~=0.18.0
python-Levenshtein
boto~=2.49.0
asyncpg~=0.21.0
gino~=1.0.1